BANG/src/main/java/Logica/Partita/Carte/Normali/BordoMarrone/Emporio.java
package Logica.Partita.Carte.Normali.BordoMarrone;

import Logica.Partita.Gestione.Giocatore;
import Logica.Partita.Gestione.Gioco;

/**
 * Questa classe crea la carta Emporio per il gioco
 */
public class Emporio extends BordoMarrone{
    /**
     * Il costruttore della carta Emporio
     * @param seme il seme della carta
     */
    public Emporio(String seme) {
        super(seme);
    }

    /**
     * Applica l'effetto della carta Emporio
     * @param gioco il gioco Bang
     * @param giocatore il giocatore che ha usato la carta
     */
    @Override
    public void ApplicaEffetto(Gioco gioco, Giocatore giocatore) {
        for (int i = 0; i < 4; i++) {
            gioco.giocatori[i].carte.add(gioco.mazzo.getFirst());
        }
    }

    /**
     * Ritorna la descrizione della carta Emporio
     * @return la descrizione della carta
     */
    public String toString() {
        return " " +  getClass().getSimpleName() + " " + getSeme() + " Girare a faccia in su tante carte del mazzo quanti sono i giocatori ancora in gioco e procedendo in senso orario, ogni giocatore sceglie una di queste carte e la mette in mano.";
    }
}
